frontend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
appdirs==1.4.4
    # via shiny
asgiref==3.8.1
    # via shiny
attrs==25.3.0
    # via environ-config
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
click==8.1.8
    # via
    #   shiny
    #   uvicorn
dotenv==0.9.9
    # via -r requirements.in
environ-config==24.1.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
htmltools==0.6.0
    # via shiny
httpcore==1.0.8
    # via httpx
httpx==0.28.1
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
linkify-it-py==2.0.3
    # via shiny
markdown-it-py==3.0.0
    # via
    #   mdit-py-plugins
    #   shiny
mdit-py-plugins==0.4.2
    # via shiny
mdurl==0.1.2
    # via markdown-it-py
msgspec==0.19.0
    # via -r requirements.in
narwhals==1.35.0
    # via shiny
orjson==3.10.16
    # via shiny
packaging==24.2
    # via
    #   htmltools
    #   shiny
prompt-toolkit==3.0.51
    # via
    #   questionary
    #   shiny
python-dotenv==1.1.0
    # via dotenv
python-multipart==0.0.20
    # via shiny
questionary==2.1.0
    # via shiny
setuptools==78.1.0
    # via shiny
shiny==1.4.0
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via shiny
typing-extensions==4.13.2
    # via
    #   anyio
    #   htmltools
    #   shiny
uc-micro-py==1.0.3
    # via linkify-it-py
uvicorn==0.34.1
    # via shiny
watchfiles==1.0.5
    # via shiny
wcwidth==0.2.13
    # via prompt-toolkit
websockets==15.0.1
    # via shiny
